server/requirements.txt
flask
flask-cors
fastapi
uvicorn[standard]
pydantic
python-dotenv
chromadb
sentence-transformers
scikit-learn
spacy
pymupdf
pytesseract
pdf2image
Pillow
numpy
tqdm
httpx
google-generativeai
werkzeug
